fix(kv-cache): saving and restoring a transposed V computes its offsets in 64 bits
With flash attention off the V cache is transposed, one row of kv_size cells per V dimension, and the state save and
restore address row j at j * kv_size. Both factors are uint32_t, so the product wrapped at 2^32:

- the save (state_write_data) also took the element size as uint32_t, so the whole byte offset was 32-bit and wrapped
  once a layer's V reached 4 GiB: an F16 V of 1024 dimensions at 2,097,152 cells wrote its row 1024 over row 0;
- the restore (state_read_data, both the contiguous and the scattered path) wrapped at 2^32 elements.

Each read or write stayed inside the tensor, so nothing failed: the saved state, or the restored cache, was silently
wrong. The served cache runs flash attention (V not transposed) and is not affected. The product is now taken in
size_t. No test: reaching the wrap takes a V of 2^31 elements or more in one layer, several GiB, beyond what CI holds.

Found by rig-lead's review of c1518d4.

+++ src/llama-kv-cache.cpp
@@ -2662,7 +2662,8 @@ void llama_kv_cache::state_write_data(llama_io_write_i & io, const cell_ranges_t
                 // Read each range of cells of v_size_el length and write out
                 for (const auto & range : cr.data) {
                     const size_t range_size = range.second - range.first;
-                    const size_t src_offset = (range.first + j * kv_size) * v_size_el;
+                    // in 64 bits: n_embd_v_gqa * kv_size can reach 2^32
+                    const size_t src_offset = (range.first + (size_t) j * kv_size) * v_size_el;
                     const size_t buf_size = range_size * v_size_el;
                     io.write_tensor(v, src_offset, buf_size);
                 }
@@ -2950,19 +2951,20 @@ bool llama_kv_cache::state_read_data(llama_io_read_i & io, uint32_t strm, uint32
                 return false;
             }
 
+            // the offsets in 64 bits: n_embd_v_gqa * kv_size can reach 2^32
             if (cell_count) {
                 if (sinfo.is_contiguous()) {
                     // Fast path: contiguous cells
                     const uint32_t h = sinfo.head();
                     for (uint32_t j = 0; j < n_embd_v_gqa; ++j) {
-                        const size_t dst_offset = (h + j * cells.size()) * v_size_el;
+                        const size_t dst_offset = (h + (size_t) j * cells.size()) * v_size_el;
                         io.read_tensor(v, dst_offset, cell_count * v_size_el);
                     }
                 } else {
                     // Slow path: scatter to non-contiguous positions, a run of cells at a time
                     for (uint32_t j = 0; j < n_embd_v_gqa; ++j) {
                         for (const auto & [c0, n] : runs) {
-                            io.read_tensor(v, (c0 + j * cells.size()) * v_size_el, n * v_size_el);
+                            io.read_tensor(v, (c0 + (size_t) j * cells.size()) * v_size_el, n * v_size_el);
                         }
                     }
                 }
